bugfix(grid): move data to mixin
